refactor(mymdc): merge mymdc config into main settings
