+++ src/zulip_write_only_proxy/settings.py
@@ -1,7 +1,10 @@
+import datetime as dt
 from pathlib import Path
 
-from pydantic import AnyUrl, DirectoryPath, SecretStr
+from pydantic import AnyUrl, DirectoryPath, HttpUrl, SecretStr
 from pydantic_settings import BaseSettings, SettingsConfigDict
+
+from .repositories import ClientRepository, ZuliprcRepository
 
 
 class Auth(BaseSettings):
@@ -10,6 +13,21 @@ class Auth(BaseSettings):
     server_metadata_url: AnyUrl = AnyUrl(
         "https://auth.exfldadev01.desy.de/application/o/zwop/.well-known/openid-configuration",
     )
+
+
+class MyMdCCredentials(BaseSettings):
+    """MyMdC Credentials. Read from `MYMDC_{key}` environment variables/`.env` file.
+
+    Get from from <https://in.xfel.eu/metadata/oauth/applications>.
+    """
+
+    id: str
+    secret: SecretStr
+    email: str
+    token_url: HttpUrl
+
+    _access_token: str = ""
+    _expires_at: dt.datetime = dt.datetime.fromisocalendar(1970, 1, 1)
 
 
 class Settings(BaseSettings):
@@ -21,10 +39,13 @@ class Settings(BaseSettings):
     config_dir: DirectoryPath = Path(__file__).parent.parent.parent / "config"
 
     auth: Auth
+    mymdc: MyMdCCredentials
+    clients: ClientRepository
+    zuliprcs: ZuliprcRepository
 
     model_config = SettingsConfigDict(
         env_prefix="ZWOP_", env_file=[".env"], env_nested_delimiter="__"
     )
 
 
-settings = Settings()  # type: ignore
+settings = Settings()  # type: ignore[call-arg]
